requirements.txt
python-dotenv
requests
beautifulsoup4
python-telegram-bot[job-queue]
tzdata
